Use exit codes of 2 and 3
so that we can distinguish between the two different ways that
pycobertura can explicitly fail and even more importantly distinguish
these types of failures from exit code 1, which is what seems to be
returned if an exception occurs.

Currently, any kind of failure returns an exit code 1 so the caller
cannot distinguish between various cases.