Refactor `reporters.py` file (#124)
We shift as much configuration and formatting into the base classes `Reporter`
and `DeltaReporter`. The main reason for this is to keep the needed changes to
this file when introducing new formats, which needs a `Reporter` as well as a
`DeltaReporter` subclass, minimal! Furthermore, it sets a standard for all the
output tables.

Possible future supportable output formats include: JSON, YAML, CSV, TSV, NSV
(new-line separated values sometimes also summarized as TSV).